Fix issue form last PR
